add tests for menus api

+++ qrodizio/views/api/menus.py
@@ -2,6 +2,7 @@ from flask import Blueprint, jsonify, abort, request
 from qrodizio.models.menus import Menu, Item
 from qrodizio.util import menus_builder
 from qrodizio.ext.database import db
+
 menus_bp = Blueprint("menus", __name__, url_prefix="/menus")
 
 
@@ -40,15 +41,16 @@ def create_menus():
     description = request.json.get("description", None)
     is_daily = request.json.get("is_daily", False)
     items = request.json.get("items")
-    
 
-
-    menu = menus_builder(name = name, description = description, is_daily = is_daily, items = items)
+    menu = menus_builder(
+        name=name, description=description, is_daily=is_daily, items=items
+    )
     menu.create()
 
-    return jsonify({"menu": menu.to_dict()}), 201
+    return jsonify({"menu": _menu_to_dict(menu)}), 201
 
-@menus_bp.route("/<int:menu_id>", methods=["PUT"])
+
+@menus_bp.route("/<int:menu_id>", methods=["POST"])
 def add_item_menu(menu_id):
     menu = Menu.query.get_or_404(menu_id)
     name = menu.name
@@ -56,25 +58,30 @@ def add_item_menu(menu_id):
     is_daily = menu.is_daily
     items = request.json.get("items")
 
-    menu = menus_builder(id = menu_id, name = name, description = description, is_daily = is_daily, items = items)
+    menu = menus_builder(
+        id=menu_id, name=name, description=description, is_daily=is_daily, items=items
+    )
     db.session.add(menu)
     db.session.commit()
 
-    return jsonify({"menu": menu.to_dict()}), 202
+    return jsonify({"menu": _menu_to_dict(menu)}), 202
+
 
 @menus_bp.route("/<int:menu_id>", methods=["PUT"])
 def edit_menu(menu_id):
     menu = Menu.query.get_or_404(menu_id)
-    name = request.json.get("name")
-    description = request.json.get("description")
-    is_daily = request.json.get("is_daily")
-    items = request.json.get("items")
+    name = request.json.get("name", None)
+    description = request.json.get("description", None)
+    is_daily = request.json.get("is_daily", False)
+    items = request.json.get("items", [])
 
-    menu = menus_builder(id = menu_id, name = name, description = description, is_daily = is_daily, items = items)
+    menu = menus_builder(
+        id=menu_id, name=name, description=description, is_daily=is_daily, items=items
+    )
     db.session.add(menu)
     db.session.commit()
 
-    return jsonify({"menu": menu.to_dict()}), 202
+    return jsonify({"menu": _menu_to_dict(menu)}), 202
 
 
 @menus_bp.route("/<int:menu_id>", methods=["DELETE"])
@@ -82,12 +89,17 @@ def delete_menu(menu_id):
     menu = Menu.query.get_or_404(menu_id)
     db.session.delete(menu)
     db.session.commit()
-    return jsonify({"sucess": "delete is working"}), 200
+    return jsonify({"success": "menu deleted"}), 202
+
 
 @menus_bp.route("/<int:menu_id>/<int:item_id>", methods=["DELETE"])
 def delete_item_menu(menu_id, item_id):
     menu = Menu.query.get_or_404(menu_id)
     item = Item.query.get_or_404(item_id)
+
+    if item not in menu.items:
+        return jsonify({"error": "Item does not belong to menu"}), 406
+
     db.session.delete(item)
     db.session.commit()
-    return jsonify({"sucess": "delete item is working"}), 200
+    return jsonify({"success": "menu item deleted"}), 202
